requirements.txt
# AI Lead Magnet Generator - Dependencies

# Web UI
gradio>=6.0.0

# LLM Client
openai>=1.0.0

# Search Provider
tavily-python>=0.3.0

# Data Validation
pydantic>=2.0.0

# Environment Configuration
python-dotenv>=1.0.0

# Python 3.10+ required
